Add files via upload
This commit contains the newest version of wormy that includes docstrings. It also includes a program that will spit out the docstrings when prompted in your local terminal. Lastly, this commit contains a file that lists the output of the docstrings. Please do let me know if you have any issues. Thank you.